Add email sending

+++ src/main/java/com/lot/crud/DTOs/ReservationDTO.java
@@ -4,13 +4,13 @@ public class ReservationDTO {
 	public Long flightNumber;
 	public Long passengerNumber;
 	public boolean didTakeoff;
-	public Integer seatNumber;
+	public Long seatNumber;
 
 	ReservationDTO() {
 
 	}
 
-	ReservationDTO(Long flightNumber, Long passengerNumber, boolean didTakeoff, Integer seatNumber) {
+	ReservationDTO(Long flightNumber, Long passengerNumber, boolean didTakeoff, Long seatNumber) {
 		this.flightNumber = flightNumber;
 		this.passengerNumber = passengerNumber;
 		this.didTakeoff = didTakeoff;
